requirements.txt
celery==4.4.0
fastapi==0.88.0
fastapi-utils==0.2.1
flower==0.9.7
kombu==4.6.7
sqlalchemy==1.4.39
uvicorn==0.20.0
vine==1.3.0
numpy==1.21.5
pandas==1.4.4
PyTrack-NTU==1.1.1
pymysql==1.0.2
pydantic==1.10.2
cryptography==37.0.1
python-dotenv==0.21.0